UPDATE: defensive code for migration

+++ database/migrations/2020_07_07_083207_add_is_table_service_flag_to_order.php
@@ -31,7 +31,9 @@ class AddIsTableServiceFlagToOrder extends Migration
     {
         if (Schema::hasTable('orders')) {
             Schema::table('orders', function (Blueprint $table) {
-                $table->dropColumn('is_table_service');
+                if (!Schema::hasColumn('orders', 'is_table_service')) {
+                    $table->dropColumn('is_table_service');
+                }
             });
         }
     }
